pyproject.toml
[project]
name = "excludarr"
version = "0.1.0"
description = "Sync Sonarr with streaming services to optimize storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pyarr>=5.2.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
]

[project.scripts]
excludarr = "excludarr.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "responses>=0.25.7",
    "respx>=0.21.1",
]
